refactor: 투표가 없는 토론게시글이 삭제할 수 있도록 수정
