+++ backend/src/test/java/com/woowacourse/gongseek/article/application/ArticleServiceTest.java
@@ -6,10 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.woowacourse.gongseek.article.domain.Article;
 import com.woowacourse.gongseek.article.domain.Category;
-import com.woowacourse.gongseek.article.domain.Content;
-import com.woowacourse.gongseek.article.domain.TempArticle;
-import com.woowacourse.gongseek.article.domain.TempTags;
-import com.woowacourse.gongseek.article.domain.Title;
 import com.woowacourse.gongseek.article.domain.repository.ArticleRepository;
 import com.woowacourse.gongseek.article.domain.repository.TempArticleRepository;
 import com.woowacourse.gongseek.article.domain.repository.dto.ArticlePreviewDto;
@@ -250,7 +246,8 @@ public class ArticleServiceTest extends IntegrationTest {
         Member notAuthorMember = memberRepository.save(new Member("judy", "juriring", "avatar.com"));
         ArticleRequest articleRequest = new ArticleRequest("질문합니다.", "내용입니다~!", Category.QUESTION.getValue(),
                 List.of("Spring"), true);
-        ArticleIdResponse savedArticle = articleService.create(new LoginMember(notAuthorMember.getId()), articleRequest);
+        ArticleIdResponse savedArticle = articleService.create(new LoginMember(notAuthorMember.getId()),
+                articleRequest);
 
         ArticleResponse articleResponse = articleService.getOne(new GuestMember(), savedArticle.getId());
 
@@ -765,22 +762,14 @@ public class ArticleServiceTest extends IntegrationTest {
         );
     }
 
-    @Transactional
     @Test
-    void 게시글을_생성하면_임시_게시글은_삭제된다() {
-        final TempArticle tempArticle = tempArticleRepository.save(TempArticle.builder()
-                .title(new Title("title"))
-                .content(new Content("content"))
-                .category(Category.QUESTION)
-                .member(member)
-                .tempTags(new TempTags(List.of("spring")))
-                .isAnonymous(false)
-                .build());
-        final ArticleRequest articleRequest = new ArticleRequest("질문합니다.", "내용입니다~!", Category.QUESTION.getValue(),
-                List.of("Spring"), true, tempArticle.getId());
+    void 투표가_없는_토론게시글을_삭제한다() {
+        Article article = articleRepository.save(
+                new Article("title2", "content2", Category.DISCUSSION, member, false));
 
-        articleService.create(new LoginMember(member.getId()), articleRequest);
+        LoginMember loginMember = new LoginMember(member.getId());
+        articleService.delete(loginMember, article.getId());
 
-        assertThat(tempArticleRepository.existsById(tempArticle.getId())).isFalse();
+        assertThat(articleRepository.findById(article.getId())).isEmpty();
     }
 }
